Refactor large files into submodules for maintainability
Split the 4 largest files that exceeded the 500-line lint limit:

- src/forges/github/client.rs (1233 → 301 lines)
  - Extract rate_limit.rs, comments.rs, labels.rs, milestones.rs, mutations.rs

- src/db/issues.rs (1218 → 284 lines)
  - Create issues/ directory with mod.rs, filter.rs, tests.rs
  - Extract row_to_issue helper to reduce duplication

- src/cli/issues.rs (969 → 103 lines)
  - Create issues/ directory with mod.rs, list.rs, write_ops.rs
  - Add parse_forge_repo and validate_jira_issue_prefix helpers to utils.rs

- src/forges/linear/client.rs (858 → 71 lines)
  - Extract auth.rs, states.rs, users.rs, issues.rs, projects.rs, comments.rs

All 149 tests pass. No functional changes.
